Harmozie Set/GetColor to VTK and MRML style
In VTK and MRML style, the color is always in double.
Also setting the \to do to the doxygen appropriate style.

+++ Modules/Loadable/Armatures/MRML/Core/vtkMRMLBoneNode.h
@@ -136,8 +136,8 @@ public:
 
   // -- Color ----------------------------------------------------------------
   /// Helper function to set/get the representation color.
-  void SetBoneColor(int rgb[3]);
-  void GetBoneColor(int rgb[3]);
+  void SetBoneColor(double rgb[3]);
+  void GetBoneColor(double rgb[3]);
 
   // -- Opacity --------------------------------------------------------------
   /// Helper function to set/get the representation color.
